Move all neighbours to test_utils module
