+++ nomos-da/network/core/src/dispersal/mod.rs
@@ -1,2 +1,5 @@
 pub mod executor;
 pub mod validator;
+
+#[cfg(test)]
+mod test {}
