+++ nomos-da/network/core/src/lib.rs
@@ -2,5 +2,7 @@ pub mod dispersal;
 pub mod protocol;
 pub mod replication;
 pub mod sampling;
+#[cfg(test)]
+pub mod test_utils;
 
 pub type SubnetworkId = u32;
