+++ nomos-da/network/core/src/replication/mod.rs
@@ -3,7 +3,6 @@ pub mod handler;
 
 #[cfg(test)]
 mod test {
-    use std::collections::HashSet;
     use std::time::Duration;
 
     use futures::StreamExt;
@@ -15,32 +14,10 @@ mod test {
     use tracing_subscriber::EnvFilter;
 
     use nomos_da_messages::common::Blob;
-    use subnetworks_assignations::MembershipHandler;
 
     use crate::replication::behaviour::{ReplicationBehaviour, ReplicationEvent};
     use crate::replication::handler::DaMessage;
-
-    #[derive(Clone)]
-    struct AllNeighbours {
-        neighbours: HashSet<PeerId>,
-    }
-
-    impl MembershipHandler for AllNeighbours {
-        type NetworkId = u32;
-        type Id = PeerId;
-
-        fn membership(&self, _self_id: &Self::Id) -> HashSet<Self::NetworkId> {
-            [0].into_iter().collect()
-        }
-
-        fn is_allowed(&self, _id: &Self::Id) -> bool {
-            true
-        }
-
-        fn members_of(&self, _network_id: &Self::NetworkId) -> HashSet<Self::Id> {
-            self.neighbours.clone()
-        }
-    }
+    use crate::test_utils::AllNeighbours;
 
     #[tokio::test]
     async fn test_connects_and_receives_replication_messages() {
